refactor: major decomposition

+++ src/severity.rs
@@ -1,4 +1,1 @@
-pub type Severity = isize;
-
-// TODO: Trait Severiry: Display.
-// as_isize() -> isize.
+pub type Severity = i32;
